Major Upgrade with Bookmarks for Series
This is a major update that refactors much of the code and adds some important features.  The Oracle database class has new methods to return the bookmark date of target time series, as well as return the row count and ping the server (for an uptime check).  The bookmark minimizes UPSERTS against the database.  The FRED series has a new "lookback" attribute to allow a window of UPSERTS since some sources revise data.